Added IcollisionHandler hpp file
